Move event procesing to job queue.
