+++ packages/uni-info-watcher/README.md
@@ -2,7 +2,52 @@
 
 ## Instructions
 
-* To start the server run `yarn server`.
+### Setup
+
+Create a postgres12 database for the job queue:
+
+```
+sudo su - postgres
+createdb uni-info-watcher-job-queue
+```
+
+Enable the `pgcrypto` extension on the job queue database (https://github.com/timgit/pg-boss/blob/master/docs/usage.md#intro).
+
+Example:
+
+```
+postgres@tesla:~$ psql -U postgres -h localhost uni-info-watcher-job-queue
+Password for user postgres:
+psql (12.7 (Ubuntu 12.7-1.pgdg18.04+1))
+SSL connection (protocol: TLSv1.3, cipher: TLS_AES_256_GCM_SHA384, bits: 256, compression: off)
+Type "help" for help.
+
+uni-watcher-job-queue=# CREATE EXTENSION pgcrypto;
+CREATE EXTENSION
+uni-info-watcher-job-queue=# exit
+```
+
+Create a postgres12 database for the uni-info watcher:
+
+```
+sudo su - postgres
+createdb uni-info-watcher
+```
+
+Update `environments/local.toml` with database connection settings for both the databases.
+
+### Run
+
+* Start the server:
+  ```bash
+  $ yarn server
+  ```
+
+* Start the job runner:
+
+  ```bash
+  $ yarn job-runner
+  ```
 
 * Run `yarn server:mock` to run server with mock data.
