+++ packages/uni-info-watcher/environments/local.toml
@@ -38,3 +38,7 @@
   [upstream.tokenWatcher]
     gqlEndpoint = "http://127.0.0.1:3001/graphql"
     gqlSubscriptionEndpoint = "http://127.0.0.1:3001/graphql"
+
+[jobQueue]
+  dbConnectionString = "postgres://postgres:postgres@localhost/uni-info-watcher-job-queue"
+  maxCompletionLag = 300
