+++ packages/uni-info-watcher/src/entity/BlockProgress.ts
@@ -1,0 +1,26 @@
+import { Entity, PrimaryGeneratedColumn, Column, Index } from 'typeorm';
+
+@Entity()
+@Index(['blockHash', 'contract'], { unique: true })
+export class BlockProgress {
+  @PrimaryGeneratedColumn()
+  id!: number;
+
+  @Column('varchar', { length: 66 })
+  blockHash!: string;
+
+  @Column('integer')
+  blockNumber!: number;
+
+  @Column('varchar', { length: 42 })
+  contract!: string;
+
+  @Column('integer')
+  numEvents!: number;
+
+  @Column('integer')
+  numProcessedEvents!: number;
+
+  @Column('boolean')
+  isComplete!: boolean
+}
