+++ packages/uni-info-watcher/src/entity/Transaction.ts
@@ -3,6 +3,8 @@ import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { decimalTransformer } from '@vulcanize/util';
 
 import { Mint } from './Mint';
+import { Burn } from './Burn';
+import { Swap } from './Swap';
 
 @Entity()
 export class Transaction {
@@ -21,6 +23,9 @@ export class Transaction {
   @OneToMany(() => Mint, mint => mint.transaction)
   mints!: Mint[];
 
-  // burns: [Burn]!
-  // swaps: [Swap]!
+  @OneToMany(() => Burn, burn => burn.transaction)
+  burns!: Burn[];
+
+  @OneToMany(() => Swap, swap => swap.transaction)
+  swaps!: Swap[];
 }
